Give pathpoint a separate file
