+++ roboteam_ai/src/control/PIDController.h
@@ -34,6 +34,7 @@ public:
     PIDController(double P, double I, double D, double timeDiff, double initial, double initial2, double prev, double prev2);
 
     bool hasInitialized();
+    bool isZero();
     void reset();
 
     void setP(double P);
