+++ roboteam_ai/src/control/positionControllers/NumTreePosControl.cpp
@@ -270,7 +270,7 @@ double NumTreePosControl::remainingStraightLinePathLength(Vector2 currentPos, Ve
 }
 
 /// create a new pathPoint using a linear acceleration ODE
-std::shared_ptr<NumTreePosControl::PathPoint> NumTreePosControl::computeNewPoint(
+std::shared_ptr<PathPoint> NumTreePosControl::computeNewPoint(
         std::shared_ptr<PathPoint> oldPoint, Vector2 subTarget) {
 
 //Copy parent
@@ -363,7 +363,7 @@ Vector2 NumTreePosControl::findCollisionPos(std::shared_ptr<PathPoint> point, do
 }
 
 /// after a collision, get new half-way targets to try to go towards
-std::pair<std::vector<Vector2>, std::shared_ptr<NumTreePosControl::PathPoint>> NumTreePosControl::getNewTargets(
+std::pair<std::vector<Vector2>, std::shared_ptr<PathPoint>> NumTreePosControl::getNewTargets(
         std::shared_ptr<PathPoint> collisionPoint) {
 
     std::shared_ptr<PathPoint> initialBranchStart = collisionPoint->backTrack(collisionPoint->t - 0.91);
@@ -389,48 +389,10 @@ std::pair<std::vector<Vector2>, std::shared_ptr<NumTreePosControl::PathPoint>> N
     return {newTargets, newBranchStart};
 }
 
-/// go back in the path until desired time or until Root
-std::shared_ptr<NumTreePosControl::PathPoint> NumTreePosControl::PathPoint::backTrack(double backTime) {
-    if (! parent)
-        return shared_from_this();
-    else if (backTime > t)
-        return shared_from_this();
-    else
-        return parent->backTrack(backTime);
-}
 
-/// go back in the path until desired collision difference or until Root
-std::shared_ptr<NumTreePosControl::PathPoint> NumTreePosControl::PathPoint::backTrack(int maxCollisionDiff) {
-    if (! parent)
-        return shared_from_this();
-    else if (maxCollisionDiff == 0)
-        return shared_from_this();
-    else
-        return parent->backTrack(collisions - parent->collisions);
-    if (collisions > parent->collisions)
-        return parent->backTrack(maxCollisionDiff - 1);
-    else
-        return parent->backTrack(maxCollisionDiff);
-}
-
-/// go back in the path until desired time, collision difference or until Root
-std::shared_ptr<NumTreePosControl::PathPoint> NumTreePosControl::PathPoint::backTrack(double backTime,
-        int maxCollisionDiff) {
-
-    if (! parent)
-        return shared_from_this();
-    else if (maxCollisionDiff == 0)
-        return shared_from_this();
-    else if (collisions > parent->collisions)
-        return parent->backTrack(maxCollisionDiff - 1);
-    else if (backTime > t)
-        return shared_from_this();
-    else
-        return parent->backTrack(backTime, maxCollisionDiff);
-}
 
 ///backTracks the path from endPoint until it hits root and outputs in order from root->endPoint
-std::vector<NumTreePosControl::PathPoint> NumTreePosControl::backTrackPath(std::shared_ptr<PathPoint> point,
+std::vector<PathPoint> NumTreePosControl::backTrackPath(std::shared_ptr<PathPoint> point,
         std::shared_ptr<PathPoint> root) {
 
     std::vector<PathPoint> path = {};
@@ -481,57 +443,6 @@ void NumTreePosControl::drawCross(Vector2 &pos, QColor color) {
 /// draw a point in the interface
 void NumTreePosControl::drawPoint(Vector2 &pos, QColor color) {
     displayData.emplace_back(pos, color);
-}
-
-/// add a child to the current path
-void NumTreePosControl::PathPoint::addChild(std::shared_ptr<PathPoint> &newChild) {
-    children.push_back(newChild);
-}
-/// add multiple children to the current path
-void NumTreePosControl::PathPoint::addChildren(std::vector<std::shared_ptr<PathPoint>> &newChildren) {
-    children.insert(children.end(), newChildren.begin(), newChildren.end());
-}
-
-/// check if a branch already has the target
-bool NumTreePosControl::PathPoint::branchHasTarget(const Vector2 &target) {
-
-    for (const auto &child : children) {
-        if ((child->currentTarget - target).length() < 0.15) {
-            return true;
-        }
-    }
-    return false;
-}
-
-/// check if ANY branch already has that target
-bool NumTreePosControl::PathPoint::anyBranchHasTarget(const Vector2 &target) {
-    auto root = backTrack(0.0);
-    return root->anyChildHasTarget(target);
-}
-
-/// check if ANY child already has that target
-bool NumTreePosControl::PathPoint::anyChildHasTarget(const Vector2 &target) {
-    for (const auto &child : children) {
-        if ((child->currentTarget - target).length() < 0.15 || child->anyChildHasTarget(target)) {
-            return true;
-        }
-    }
-    return false;
-}
-
-/// check if ANY parent already has that target
-bool NumTreePosControl::PathPoint::anyParentHasTarget(const Vector2 &target) {
-    if (parent) {
-        if ((parent->currentTarget - target).length() < 0.15 || parent->anyParentHasTarget(target)) {
-            return true;
-        }
-    }
-    return false;
-}
-
-/// check if a collision is occuring
-bool NumTreePosControl::PathPoint::isCollision(Vector2 target, double distance) {
-    return (target - pos).length() < distance;
 }
 
 }// control
