+++ roboteam_ai/src/skills/AvoidBall.h
@@ -22,7 +22,7 @@ class AvoidBall : public Skill {
         explicit AvoidBall(std::string name, bt::Blackboard::Ptr blackboard = nullptr);
         bt::Node::Status onUpdate() override;
     private:
-        control::PositionController gtp;
+        control::PositionManager gtp;
         enum Progression {
           RUNNING, DONE, FAIL
         };
