+++ roboteam_ai/src/skills/DefendOnRobot.h
@@ -5,7 +5,7 @@
 #ifndef ROBOTEAM_AI_DEFENDONROBOT_H
 #define ROBOTEAM_AI_DEFENDONROBOT_H
 
-#include <roboteam_ai/src/control/PositionController.h>
+#include <roboteam_ai/src/control/PositionManager.h>
 #include "Skill.h"
 
 namespace rtt {
@@ -13,7 +13,7 @@ namespace ai {
 
 class DefendOnRobot : public Skill {
     private:
-        control::PositionController goToPos;
+        control::PositionManager goToPos;
         double angleBetweenRobots;
         Vector2 newPosition;
         int pickOpponentToCover();
