+++ roboteam_ai/src/skills/GTPSpecial.h
@@ -5,7 +5,7 @@
 #ifndef ROBOTEAM_AI_BASICGOTOPOS_H
 #define ROBOTEAM_AI_BASICGOTOPOS_H
 
-#include <roboteam_ai/src/control/PositionController.h>
+#include <roboteam_ai/src/control/PositionManager.h>
 #include <roboteam_ai/src/coach/GeneralPositionCoach.h>
 #include <roboteam_ai/src/coach/Ballplacement.h>
 #include <roboteam_ai/src/utilities/Field.h>
